Ajax requests by path
With this update, ajax requests should now be made by reference to the
path to the ui block, rather than just the id of the target block you
are calling the method on. This way blocks that are not relevant to the
call don’t need to be loaded at all.
www.example.com/?ui=block1.block2.targetBlock&ajax=targetMethod